refactored categories to use user id from session
