+++ README.md
@@ -31,9 +31,10 @@
 #### Categories
 
 - GET `/api/categories` returns all categories
-- GET `http://localhost:3002/api/categories/get_categories_by_id/:id` returns categories for a specific user
-- DELETE `/api/categories/:id/delete` deletes category by id
+- GET `http://localhost:3002/api/categories/get_categories_by_id/` returns categories for a specific user
+- DELETE `/api/categories/delete` deletes category by id
   - assumes user is already logged in as users should only be able to delete categories they own
+  - payload from front end, needs to include cat_id
 - POST `/api/categories` adds a new category
   payload:
