+++ routes/authCheck.js
